These are mostly optimizations from https://github.com/robur-coop/miragevpn/pull/223 but also some new ones since we have encrypt_into etc. functions.

`main-bench.exe` is built from 6099a4bd4134eedddd16090a083dc58d246eeae1 (unfortunately without --release)
![image](https://github.com/user-attachments/assets/57fb1a27-a6e5-4fce-bf6b-bcb461336b1c)

As you can see above there are improvements in execution time, but most notably the number of minor words allocations is almost halved. When testing against an openvpn server I didn't observe much of a difference except, curiously, a slowdown in upload speed (from 826 Mb/s to 757 Mb/s). Why that would I don't know.