@reynir 2024-09-27
> When testing against an openvpn server I didn't observe much of a difference except, curiously, a slowdown in upload speed (from 826 Mb/s to 757 Mb/s). Why that would I don't know.

Ah! This was with the UDP configuration. With TCP there is a slight speedup (down: 919 Mb/s -> 1.07 Gb/s, up: 599 Mb/s -> 652 Mb/s)