@reynir 2024-09-28
I accidentally committed and pushed 15ba1eda481228c8560eacd8b7a12e825d271af2 and eba6e587fac9b1756b63e92e13e327f546208e24 to `main`. They were a fix to the failing e2e test where I observed locally that it failed due to a Data_v1 packet with key id 0 arriving seemingly just after a successful rekey for key id 1. I don't know if it's the correct thing to do, but to me it seems for data packets we could be more lenient and just drop them.