@reynir 2024-09-28
Ah wait the e2e test already fails at the static client configuration. Locally, it initially also failed there for me, but that was due to firewall rules.