@reynir 2024-09-30
I cleaned up the git history a bit by merging two commits and dropping the CI debugging commit.

It fails on OCaml 5.3~alpha1 due to `effect` suddenly being a keyword. I'll address this separately.